Ported fixes from 110f3859c185762b5afd2f99e8c8840c1ebed6e0

+++ src/utils.h
@@ -10,7 +10,7 @@
 // CAN Transmit Error Counter Register
 #define TECR 0x4005084F
 
-float msg_duration_micros(const uint32_t bps, const uint8_t dlc);
+uint64_t msg_duration_micros(const uint32_t bps, const uint8_t dlc);
 
 template<typename T>
 T avg(const T data[], const unsigned int len) {
